Fix/random (#160)
* Correct constant definition in safe.go

Updated the `_intSize` variable to be a constant in safe.go. This ensures the code is more robust and adheres to the best practice of using constants for values that aren't supposed to change at runtime.

* Update `randBytes` constant in random.go

The `randBytes` constant in random.go has been updated to include numbers 0 and 1. This change extends the range of random values that can be generated, ensuring more randomness and security.