Fix/coarsetime (#162)
* Refactor TruncateHour and TruncateDay functions

The TruncateHour and TruncateDay functions in truncate.go were updated to use the current date and location. This could ensure more accurate results as the new implementation guarantees to truncate relevant parts, excluding hours for TruncateDay and minutes and seconds for TruncateHour.

* Fix TruncateWeekday logic in truncate.go

Updated the return value of TruncateWeekday function in truncate.go to subtract the weekday instead of adding. This change correctly aligns the TruncateWeekday function with how the start of the week is typically calculated.