Add new test case to exclude_test.go
A new test case has been added to exclude_test.go in the cidr package. This additional test is designed to provide a more in-depth evaluation of the IP range exclusion-related behavior of the DelAddress method. The enhancement not only increases overall test coverage but also aids in identifying potential regressions in the future.